--- imgs-20221021T020631Z-001/imgs/pygame.py
@@ -1,228 +1,0 @@
-import pygame
-import os
-import random
-
-largura_tela = 500
-altura_tela = 800
-
-imagem_cano = pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'pipe.png')))
-imagem_chao = pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'base.png')))
-imagem_plano_fundo = pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'bg.png')))
-imagens_passaro = [
-    pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'bird1.png'))),
-    pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'bird2.png'))),
-    pygame.transform.scale2x(pygame.image.load(os.path.join('imgs', 'bird3.png'))),
-]
-
-pygame.font.init()
-fonte_pontos = pygame.font.SysFont('arial', 50)
-
-class Passaro:
-    imgs = imagens_passaro
-    # Animação rotação
-    rotacao_maxima = 25
-    velocidade_rotacao = 20
-    tempo_animacao = 5
-
-    def __init__(self, x, y):
-        self.x = x
-        self.y = y
-        self.angulo = 0
-        self.velocidade = 0
-        self.altura = self.y
-        self.tempo = 0
-        self.contagem_imagem = 0
-        self.imagem = imgs[0]
-
-    def pular(self):
-        self.velocidade = -10.5
-        self.tempo = 0
-        self.altura = self.y
-    
-    def mover(self):
-        #calcular deslocamento
-        self.tempo += 1
-        deslocamento = 1.5 * (self.tempo**2) + self.velocidade * self.tempo
-        
-        #restringir deslocamento
-        if deslocamento > 16:
-            deslocamento = 16
-        elif deslocamento < 0:
-            deslocamento -= 2
-            self.y += deslocamento
-        #angulo passaro
-        if deslocamento < 0 or self.y < (self.altura + 50):
-            if self.angulo < self.rotacao_maxima:
-                self.angulo = self.rotacao_maxima
-        else: 
-            if self.angulo > -90:
-                self.angulo -= self.velocidade_rotacao
-    def desenho(self):
-        # escolha imagem passaro
-        self.contagem_imagem += 1
-
-        if self.contagem_imagem < self.tempo_animacao:
-            self.imagem = self.imgs[0]
-        elif self.contagem_imagem < self.tempo_animacao * 2:
-            self.imagem = self.imgs[1]
-        elif self.contagem_imagem < self.tempo_animacao * 2:
-            self.imagem = self.imgs[2]
-        elif self.contagem_imagem < self.tempo_animacao * 4:
-            self.imagem = self.imgs[1]
-        elif self.contagem_imagem >= self.tempo_animacao * 4 + 1:
-            self.imagem = self.imgs[0]
-            self.contagem_imagem = 0
-
-        # escolha passaro caindo
-        if self.angulo <= -80:
-            self.imagem = self.imgs[1]
-            self.contagem_imagem = self.tempo_animacao * 2
-
-        # desenho imagem passaro
-        imagem_ratacionada = pygame.transform.rotate(self.imagem, self.angulo)
-        pos_centro_img = self.imagem.get_rect(topleft = (self.x, self.y)).center
-        retangulo = imagem_ratacionada.get_rect(center = pos_centro_img)
-        tela.blit(imagem_ratacionada, retangulo.topleft)
-    def get_mask(self):
-        return pygame.mask.from_surface(self.imagem)
-
-class Cano:
-    distancia = 200
-    velocidade = 5
-    
-    def __init__(self, x):
-        self.x = x
-        self.altura = 0
-        self.posicao_topo = 0
-        self.posicao_base = 0
-        self.cano_topo = pygame.transform.flip(imagem_cano, False, True)
-        self.cano_base = imagem_cano
-        self.passou = False
-        self.definir_altura()
-
-    def definir_altura(self):
-        self.altura = random.randrange(50, 450)
-        self.posicao_topo = self.altura - self.cano_topo.get_height()
-        self.posicao_base = self.altura + self.distancia
-    
-    def mover(self):
-        self.x1 -= self.velocidade
-        self.x2 -= self.velocidade
-
-        if self.x1 + self.largura < 0:
-            self.x1 = self.x2 + self.largura
-        if self.x2 + self.largura < 0:
-            self.x2 = self.x1 + self.largura
-    
-    def desenhar(self, tela):
-        tela.blit(self.cano_topo, (self.x, self.posicao_topo))
-        tela.blit(self.cano_base, (self.x, self.posicao_base))
-    
-    def colidir(self, passaro):
-        passaro_mask = passaro.get_mask()
-        topo_mask = pygame.mask.from_surface(self.cano_topo)
-        base_mask = pygame.mask.from_surface(self.cano_base)
-
-        distancia_topo = (self.x - passaro.x, self.posicao_topo - round(passaro.y))
-        distancia_base = (self.x - passaro.x, self.posicao_base - round(passaro.y))
-
-        base_ponto = passaro_mask.overlap(base_mask, distancia_base)
-        topo_ponto = passaro_mask.overlap(base_mask, distancia_topo)
-
-        if base_ponto or topo_ponto:
-            return True
-        else:
-            return False
-
-class Chao:
-    velocidade = 5
-    largura = imagem_chao.get_width()
-    imagem = imagem_chao
-    
-    def __init__(self, y):
-        self.y = y
-        self.x0 = 0
-        self.x1 = self.largura
-
-    def mover(self):
-        self.x1 -= self.velocidade
-        self.x2 -= self.velocidade
-
-        if self.x1 + self.largura < 0:
-            self.x1 = self.x1 + self.largura
-
-        if self.x2 + self.largura < 0:
-            self.x2 = self.x2 + self.largura
-    def desenhar(self, tela):
-        tela.blit(self.imagem, (self.x1, self.y))
-        tela.blit(self.imagem, (self.x2, self.y))
-
-def desenhar_tela(tela, passaros, canos, chao, pontos):
-    tela.blit(imagem_plano_fundo,(0,0))
-
-    for passaro in passaros:
-        passaro.desenhar(tela)
-
-    for cano in canos:
-        cano.desenhar(tela)
-    
-    texto = fonte_pontos.render(f,"Pontuação:  {pontos}", 1, (255, 255, 255))
-    tela.blit(texto, (largura_tela - 10 - texto.get_width(), 10))
-    chao.desenhar(tela)
-    pygame.display.update()
-
-def main():
-    passaros = [passaros(230, 350)]
-    chao = Chao(730)
-    canos = [Cano(700)]
-    tela = pygame.display.set_mode(largura_tela, altura_tela)
-    pontos = 0
-    relogio = pygame.time.clock()
-
-    rodando = True
-    while rodando:
-
-        relogio.tick(30)
-        
-        for evento in pygame.event.get():
-            if evento.type == pygame.QUIT:
-                rodando = False
-                pygame.quit()
-                quit()
-            if evento.type == pygame.KEYDOWN:
-                if evento.key == pygame.K_SPACE:
-                    for passaro in passaros:
-                        passaro.pular()
-        #mover cenário
-        for passaro in passaros:
-            passaro.mover()
-        chao.mover()
-
-        adicionar_cano = False
-        remover_canos = []
-        for cano in canos:
-            for i, passaro in enumerate(passaros):
-                if cano.colidir(passaro):
-                    passaros.pop(i)
-                if not cano.passou and passaro.x > cano.x:
-                    cano.passou = True
-                    adicionar_cano = True
-                cano.mover()
-                if cano.x + cano.cano_topo.get_width() < 0:
-                    remover_cano.appdend(cano)
-
-            if adicionar_cano:
-                pontos += 1
-                canos.append(cano(600))
-
-            for cano in remover_canos:
-                canos.remove(cano)
-            
-            for i, passaro in enumerate(passaros):
-                if (passaro.y + passaro.imagem.get_height()) > chao.y or passaro.y < 0:
-                    passaros.pop(i)
-
-            desenhar_tela(tela, passaros, canos, chao, pontos)
-
-if __name__ == "__main__":
-    main()
