--- pygame.py
@@ -1,11 +1,0 @@
-import pygame
-import os
-import random
-
-largura_tela = 500
-altura_tela = 800
-
-imagem_cano = pygame.image.load()
-imagem_chao = 
-imagem_plano_fundo = 
-imagem_passaro =
